Complete UI Overhaul

+++ app/src/main/java/com/example/finalproject/ui/Fragments/Fragment_usuario_reservados.java
@@ -1,5 +1,6 @@
-package com.example.finalproject.ui;
+package com.example.finalproject.ui.Fragments;
 
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
@@ -7,15 +8,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.ImageView;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
+import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
@@ -24,6 +20,7 @@ import com.example.finalproject.R;
 import com.example.finalproject.models.Reservation;
 import com.example.finalproject.network.ReservationApiService;
 import com.example.finalproject.network.RetrofitInstance;
+import com.example.finalproject.ui.Activities.Pantalla_admin_reservados;
 import com.example.finalproject.ui.Activities.Pantalla_usuario_QR;
 import com.example.finalproject.utils.SessionDataManager;
 
@@ -33,103 +30,67 @@ import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class Pantalla_usuario_reservados extends AppCompatActivity {
+public class Fragment_usuario_reservados extends Fragment {
 
     private RecyclerView recyclerView;
     private ReservationAdapter adapter;
 
+    public Fragment_usuario_reservados() {
+        // Required empty public constructor
+    }
+
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        EdgeToEdge.enable(this);
-        setContentView(R.layout.pantalla_usuario_reservados);
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        View view = inflater.inflate(R.layout.fragment_usuario_reservados, container, false);
+        Button btnVerTodasReservas = view.findViewById(R.id.btnVerTodasReservas);
 
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
-            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
-            return insets;
-        });
-        ImageView imageView1 = findViewById(R.id.homeIcon2);
-        imageView1.setOnClickListener(v -> {
-            Intent intent = new Intent(Pantalla_usuario_reservados.this, Pantalla_usuario_inicial.class);
-            startActivity(intent);
-        });
-        ImageView imageView2 = findViewById(R.id.imageView2);
-        imageView2.setOnClickListener(v -> {
-            Intent intent = new Intent(Pantalla_usuario_reservados.this, Pantalla_usuario_info.class);
-            startActivity(intent);
-        });
-
-        recyclerView = findViewById(R.id.recyclerView);
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        // Verificar si el usuario es ADMIN
+        String role = SessionDataManager.getInstance().getCurrentUser().getRole();
+        if ("ADMIN".equalsIgnoreCase(role)) {
+            // Si es admin, activar el botón
+            btnVerTodasReservas.setVisibility(View.VISIBLE);
+            btnVerTodasReservas.setEnabled(true);
+            btnVerTodasReservas.setClickable(true);
+            btnVerTodasReservas.setFocusable(true);
+            btnVerTodasReservas.setOnClickListener(v -> {
+                Intent intent = new Intent(requireContext(), Pantalla_admin_reservados.class);
+                startActivity(intent);
+            });
+        }
+        recyclerView = view.findViewById(R.id.recyclerView);
+        recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         DividerItemDecoration divider = new DividerItemDecoration(
                 recyclerView.getContext(), LinearLayoutManager.VERTICAL);
         recyclerView.addItemDecoration(divider);
-
-
-        if ("ADMIN".equalsIgnoreCase(SessionDataManager.getInstance().getCurrentUser().getRole())) {
-            fetchAllReservations();
-        } else {
-            fetchReservationsByUid(SessionDataManager.getInstance().getCurrentUser().getId());
-        }
-    }
-
-
-    private void fetchAllReservations() {
-        ReservationApiService apiService = RetrofitInstance.getRetrofitInstance(this).create(ReservationApiService.class);
-        Call<List<Reservation>> call = apiService.getAllReservations();  // Usamos el endpoint GET sin filtro
-
-        call.enqueue(new Callback<List<Reservation>>() {
-            @Override
-            public void onResponse(Call<List<Reservation>> call, Response<List<Reservation>> response) {
-                if (response.isSuccessful() && response.body() != null) {
-                    List<Reservation> reservations = response.body();
-                    if (reservations.isEmpty()) {
-                        Toast.makeText(Pantalla_usuario_reservados.this, "No hay reservas disponibles", Toast.LENGTH_SHORT).show();
-                    } else {
-                        adapter = new ReservationAdapter(reservations);  // Creas el adaptador con las reservas recibidas
-                        recyclerView.setAdapter(adapter);  // Asignamos el adaptador al RecyclerView
-                    }
-                } else {
-                    Log.e("API", "Error en la respuesta: " + response.code());
-                    // Puedes mostrar un mensaje o manejar el error aquí
-                }
-            }
-
-            @Override
-            public void onFailure(Call<List<Reservation>> call, Throwable t) {
-                Log.e("API", "Error de red: ", t);
-                // Mostrar un mensaje de error indicando que hubo un problema con la conexión
-            }
-        });
+        fetchReservationsByUid(SessionDataManager.getInstance().getCurrentUser().getId());
+        return view;
     }
 
     private void fetchReservationsByUid(int id) {
-        ReservationApiService apiService = RetrofitInstance.getRetrofitInstance(this).create(ReservationApiService.class);
+        ReservationApiService apiService = RetrofitInstance.getRetrofitInstance(getActivity()).create(ReservationApiService.class);
         Call<List<Reservation>> call = apiService.getAllReservationsByUserId(id);
 
         call.enqueue(new Callback<List<Reservation>>() {
             @Override
             public void onResponse(Call<List<Reservation>> call, Response<List<Reservation>> response) {
                 if (response.isSuccessful() && response.body() != null) {
-                    // Aquí se puede manejar la respuesta y actualizar la UI (por ejemplo, RecyclerView)
                     List<Reservation> reservations = response.body();
                     if (reservations.isEmpty()) {
-                        Toast.makeText(Pantalla_usuario_reservados.this, "No hay reservas disponibles", Toast.LENGTH_SHORT).show();
+                        Toast.makeText(getActivity(), "No hay reservas disponibles", Toast.LENGTH_SHORT).show();
                     } else {
-                        adapter = new ReservationAdapter(reservations);  // Creas el adaptador con las reservas recibidas
-                        recyclerView.setAdapter(adapter);  // Asignamos el adaptador al RecyclerView
+                        adapter = new ReservationAdapter(reservations);
+                        recyclerView.setAdapter(adapter);
                     }
                 } else {
                     Log.e("API", "Error en la respuesta: " + response.code());
-                    // Aquí puedes mostrar un mensaje o manejar el error
                 }
             }
 
             @Override
             public void onFailure(Call<List<Reservation>> call, Throwable t) {
                 Log.e("API", "Error de red: ", t);
-                // Aquí puedes mostrar un mensaje indicando que hubo un error en la conexión
             }
         });
     }
@@ -172,7 +133,6 @@ public class Pantalla_usuario_reservados extends AppCompatActivity {
             }
 
             void bind(Reservation reserva) {
-                // Formateo de fecha y hora
                 String rawStart = reserva.getStartTime();
                 String rawEnd = reserva.getEndTime();
 
@@ -180,17 +140,23 @@ public class Pantalla_usuario_reservados extends AppCompatActivity {
                 String horaInicio = rawStart.substring(11, 16);
                 String horaFin = rawEnd.substring(11, 16);
 
-                // Asignar valores a los TextViews
                 classroomId.setText("Aula: " + reserva.getClassroomId());
                 date.setText("Fecha: " + fecha);
                 timeRange.setText("Horario: Inicio " + horaInicio + " - Fin " + horaFin);
 
-                //btnDelete.setVisibility(View.VISIBLE);
                 btnDelete.setOnClickListener(v -> {
-                    eliminarReserva(reserva.getId(), getAdapterPosition());
+                    new AlertDialog.Builder(getContext())
+                            .setTitle("Confirmación")
+                            .setMessage("¿Estás seguro de que quieres borrar la Reserva? Esto no se puede deshacer")
+                            .setPositiveButton("Sí", (dialog, which) -> {
+                                eliminarReserva(reserva.getId(), getAdapterPosition());
+                            })
+                            .setNegativeButton("No", (dialog, which) -> dialog.dismiss())
+                            .show();
                 });
+
                 btnShowQR.setOnClickListener(v -> {
-                    Intent intent = new Intent(Pantalla_usuario_reservados.this, Pantalla_usuario_QR.class);
+                    Intent intent = new Intent(getActivity(), Pantalla_usuario_QR.class);
                     intent.putExtra("reservation", reserva);
                     startActivity(intent);
                 });
@@ -198,7 +164,7 @@ public class Pantalla_usuario_reservados extends AppCompatActivity {
         }
 
         private void eliminarReserva(int reservaId, int position) {
-            ReservationApiService apiService = RetrofitInstance.getRetrofitInstance(Pantalla_usuario_reservados.this)
+            ReservationApiService apiService = RetrofitInstance.getRetrofitInstance(getActivity())
                     .create(ReservationApiService.class);
 
             Call<Void> call = apiService.deleteReservation(reservaId);
